drivers: npmx: shell: move led code to the separate file
